backend/requirements.txt
django
djangorestframework
drf-yasg


coreapi
coreschema


psycopg2-binary
django-allauth
django-cors-headers
python-decouple
